examples: refactor diffusion generation
